backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
certifi==2026.4.22
click==8.3.3
colorama==0.4.6
ecdsa==0.19.2
eval_type_backport==0.3.1
fastapi==0.136.1
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
importlib_metadata==8.7.1
jsonpath-python==1.1.5
mistralai==2.4.2
opentelemetry-api==1.39.1
opentelemetry-semantic-conventions==0.60b1
passlib==1.7.4
pyasn1==0.6.3
pydantic==2.13.3
pydantic_core==2.46.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
zipp==3.23.1
